a committ that may make trouble,
I changed the virtuemart_userinfo_id to a normal increment. 

I fixed the vendor slug, was pointing to the normal table and the slug is in the language table, which is not working

TODO checking how old virtuemart_userinfo_id behave, when changing to autoincrement id, maybe we must add it to the importer.

Small adjustments and little fixes to the mediahandler/uploader 

git-svn-id: http://dev.virtuemart.net/svn/virtuemart/trunk@5245 87cc427b-4a1e-0410-b729-b2260ec0f063